refactor: move handler functions closer to server script
