--- aiengine/runtime/neurdbrt/app/handlers/__init__.py
@@ -1,4 +1,0 @@
-from .finetune import *
-from .inference import *
-from .setup import *
-from .train import *
