+++ aiengine/runtime/neurdbrt/app/setup.py
@@ -4,7 +4,7 @@ from typing import List, Optional, Tuple
 
 import numpy as np
 from neurdbrt.config import DEVICE
-from neurdbrt.dataloader.stream_libsvm_dataset import StreamingDataSet
+from neurdbrt.dataloader import StreamingDataSet
 from neurdbrt.model import build_model
 from neurdbrt.repo import ModelRepository
